feat: file delete 로직 개발 및 deleteAll 메서드 트랜잭션 처리

+++ src/main/java/jaeseok/numble/mybox/folder/service/FolderService.java
@@ -33,7 +33,7 @@ public class FolderService {
         Folder folder = folderRepository.findById(id)
                 .orElseThrow(() -> new MyBoxException(ResponseCode.FOLDER_NOT_FOUND));
 
-        int fileCount = fileService.deleteAll(folder.getAllChildFiles());
+        Long fileCount = fileService.deleteAll(folder.getAllChildFiles());
 
         folderRepository.delete(folder);
